+++ alacritty/src/window.rs
@@ -47,6 +47,7 @@ use x11_dl::xlib::{Display as XDisplay, PropModeReplace, XErrorEvent, Xlib};
 use alacritty_terminal::config::Colors;
 use alacritty_terminal::config::{Decorations, StartupMode, WindowConfig};
 use alacritty_terminal::event::Event;
+use alacritty_terminal::term::color::Rgb;
 #[cfg(not(windows))]
 use alacritty_terminal::term::{SizeInfo, Term};
 
@@ -202,6 +203,14 @@ impl Window {
                 if let Some(parent_window_id) = config.window.embed {
                     x_embed_window(windowed_context.window(), parent_window_id);
                 }
+                // Perform glClear() on-screen to avoid drawing color over on the window
+                // at incorrect sizes.
+                windowed_context.window().set_visible(true);
+                gl_clear(config.colors.primary.background, config.background_opacity());
+                // We make these GL calls here directly, instead of going through QuadRenderer,
+                // to avoid having to initialize SizeInfo. Also, swap buffers to avoid glitches.
+                windowed_context.swap_buffers().expect("swap buffers");
+                gl_finish();
             } else {
                 // Apply client side decorations theme.
                 let theme = AlacrittyWaylandTheme::new(&config.colors);
@@ -480,4 +489,25 @@ fn x_embed_window(window: &GlutinWindow, parent_id: c_ulong) {
 unsafe extern "C" fn xembed_error_handler(_: *mut XDisplay, _: *mut XErrorEvent) -> i32 {
     error!("Could not embed into specified window.");
     std::process::exit(1);
+}
+
+#[inline]
+#[cfg(not(any(target_os = "macos", windows)))]
+fn gl_finish() {
+    unsafe {
+        gl::Finish();
+    }
+}
+
+#[inline]
+fn gl_clear(color: Rgb, alpha: f32) {
+    unsafe {
+        gl::ClearColor(
+            (f32::from(color.r) / 255.0).min(1.0) * alpha,
+            (f32::from(color.g) / 255.0).min(1.0) * alpha,
+            (f32::from(color.b) / 255.0).min(1.0) * alpha,
+            alpha,
+        );
+        gl::Clear(gl::COLOR_BUFFER_BIT);
+    }
 }